web/apps/requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.1.0
boto3==1.28.29
botocore==1.31.29
celery==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coloredlogs==15.0.1
cron-descriptor==1.4.0
Django==4.2.3
django-appconf==1.0.5
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-compressor==4.4
django-filter==23.2
django-storages==1.13.2
django-timezone-field==6.0
djangorestframework==3.14.0
flatbuffers==23.5.26
flower==2.0.1
humanfriendly==10.0
humanize==4.8.0
idna==3.4
imageio==2.31.1
jmespath==1.0.1
kombu==5.3.1
lazy_loader==0.3
llvmlite==0.40.1
mpmath==1.3.0
networkx==3.1
numba==0.57.1
numpy==1.24.4
onnxruntime==1.15.1
opencv-python-headless==4.8.0.76
packaging==23.1
Pillow==10.0.0
platformdirs==3.10.0
pooch==1.7.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.24.0
psycopg2==2.9.7
PyMatting==1.1.8
python-crontab==3.0.0
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.4.1
rcssmin==1.1.1
redis==5.0.0
rembg==2.0.50
requests==2.31.0
rjsmin==1.2.1
s3transfer==0.6.2
scikit-image==0.21.0
scipy==1.11.2
six==1.16.0
sqlparse==0.4.4
sympy==1.12
tifffile==2023.8.12
tornado==6.3.3
tqdm==4.66.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
vine==5.0.0
wcwidth==0.2.6
